README.md
QR Code Generator is a simple Node.js application that allows users to generate QR codes for any URL they provide. By using the inquirer package, the application prompts the user to enter a URL and then generates a corresponding QR code. This QR code is saved as an image file (qr_img.png), and the provided URL is also saved in a text file (URL.txt). The project is designed for ease of use and can be run in a Node.js environment with minimal setup.